# Fix gcc8 compilation error