… src/shogun/mathematics/linalg/backend/eigen/EigenSolvers.cpp