This fix is for issue #4419. The solution seems to work for g++ (Ubuntu 8.2.0-7ubuntu1) 8.2.0 but it might need some tinkering.