@gf712 2019-02-04
Once we reach C++17 this will be much simpler as we can just dispatch with `if constexpr(...)`. Until then I think you need to do SFINAE, since gcc8 doesn't like this anymore. You shouldn't do partial specialisations of class members.
It's not ideal but you could go around it by adding a `std::enable_if_t<std::is_same<T, float64_t>::value>* = nullptr` and `std::enable_if_t<!std::is_same<T, float64_t>::value>* = nullptr` for each case and then it should compile fine.