@avramidis 2019-02-04
Thanks! I will try this solution.