@gf712 2019-02-04
Nevermind, you end up with the same linker error. I am not sure how to fix it... you might have to just move the stuff from the translation unit to the header?