@avramidis 2019-02-04
OK. I need to build shogun on windows somehow and try to fix the problem there.