@avramidis 2019-02-05
OK, seems to work fine now. With these changes I can build with gcc8 in Ubuntu 18.10.