@avramidis 2019-02-09
Benchmarks results with #4519 

**Before gcc8 fix  with gcc7**

Start 1: RandomFourierDotFeatures_benchmark
1/6 Test 1: RandomFourierDotFeatures_benchmark ...   Passed   22.11 sec
    Start 2: HashedDocDotFeatures_benchmark
2/6 Test 2: HashedDocDotFeatures_benchmark .......   Passed   13.90 sec
    Start 3: RefCount_benchmark
3/6 Test 3: RefCount_benchmark ...................   Passed    0.69 sec
    Start 4: BasicOps_benchmark
4/6 Test 4: BasicOps_benchmark ...................   Passed   37.75 sec
    Start 5: Misc_benchmark
5/6 Test 5: Misc_benchmark .......................   Passed   35.13 sec
    Start 6: SGMatrix_benchmark
6/6 Test 6: SGMatrix_benchmark ...................   Passed   17.60 sec

100% tests passed, 0 tests failed out of 6

Label Time Summary:
benchmark    = 127.19 sec*proc (6 tests)

Total Test time (real) = 127.20 sec

**After gcc8 fix with gcc7**

Start 1: RandomFourierDotFeatures_benchmark
1/6 Test 1: RandomFourierDotFeatures_benchmark ...   Passed   21.54 sec
    Start 2: HashedDocDotFeatures_benchmark
2/6 Test 2: HashedDocDotFeatures_benchmark .......   Passed   13.65 sec
    Start 3: RefCount_benchmark
3/6 Test 3: RefCount_benchmark ...................   Passed    0.69 sec
    Start 4: BasicOps_benchmark
4/6 Test 4: BasicOps_benchmark ...................   Passed   37.67 sec
    Start 5: Misc_benchmark
5/6 Test 5: Misc_benchmark .......................   Passed   34.92 sec
    Start 6: SGMatrix_benchmark
6/6 Test 6: SGMatrix_benchmark ...................   Passed   16.87 sec

100% tests passed, 0 tests failed out of 6

Label Time Summary:
benchmark    = 125.34 sec*proc (6 tests)

Total Test time (real) = 125.35 sec